backend/requirements.txt
APIFlask==1.1.0
apispec==5.2.2
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
Flask==2.1.2
Flask-HTTPAuth==4.7.0
Flask-JWT-Extended==4.4.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
marshmallow==3.17.0
packaging==21.3
pycparser==2.21
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.39
webargs==8.1.0
Werkzeug==2.1.2
zipp==3.8.0
google-search-results
fuzzywuzzy
pandas
tencentcloud-sdk-python
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
yake
nltk